+++ test/Test.Shared/TestAssertionException.cs
@@ -1,0 +1,11 @@
+namespace Test.Shared
+{
+    using System;
+
+    public class TestAssertionException : Exception
+    {
+        public TestAssertionException(string message) : base(message)
+        {
+        }
+    }
+}